test: cover patient add switch delete
